+++ java/src/main/java/vowpalWabbit/learner/VWMulticlassLearner.java
@@ -15,5 +15,5 @@ final public class VWMulticlassLearner extends VWIntLearner {
     protected native int predictMultiline(String[] example, boolean learn, long nativePointer);
 
     @Override
-    protected native float[] rawPredict(String example, boolean learn, long nativePointer);
+    protected native double[] rawPredict(String example, boolean learn, long nativePointer);
 }
